+++ lib/mobility/attributes.rb
@@ -156,10 +156,6 @@ with other backends.
       def mobility_attribute?(name)
         mobility_attributes.include?(name.to_s)
       end
-
-      # @!method translated_attribute_names
-      # @return (see #mobility_attributes)
-      alias translated_attribute_names mobility_attributes
     end
   end
 end